+++ src/lang/en.js
@@ -232,6 +232,11 @@ export default {
     passwordResetTokenCreated: 'Password reset token was created',
     accountCreated: 'New account was created!'
   },
+  statuses: {
+    statuses: 'Statuses',
+    instanceFilter: 'Instance filter',
+    loadMore: 'Load more'
+  },
   userProfile: {
     tags: 'Tags',
     moderator: 'Moderator',